refactor: moved RequestHandler creation to UdpServer::start()
