+++ src/udp/server.rs
@@ -7,7 +7,9 @@ use log::{debug, info};
 use tokio::net::UdpSocket;
 
 use crate::tracker::tracker::TorrentTracker;
-use crate::udp::{handle_packet, MAX_PACKET_SIZE};
+use crate::udp::{MAX_PACKET_SIZE};
+use crate::udp::connection::secret::Secret;
+use crate::udp::packet_handler::RequestHandler;
 
 pub struct UdpServer {
     socket: Arc<UdpSocket>,
@@ -25,11 +27,16 @@ impl UdpServer {
     }
 
     pub async fn start(&self) {
+        let server_secret = Secret::new([0;32]);
+        let request_handler = Arc::new(RequestHandler::new(server_secret));
+
         loop {
             let mut data = [0; MAX_PACKET_SIZE];
             let socket = self.socket.clone();
             let tracker = self.tracker.clone();
+            let request_handler = request_handler.clone();
 
+            // needed for graceful shutdown
             tokio::select! {
                 _ = tokio::signal::ctrl_c() => {
                     info!("Stopping UDP server: {}..", socket.local_addr().unwrap());
@@ -41,7 +48,7 @@ impl UdpServer {
                     debug!("Received {} bytes from {}", payload.len(), remote_addr);
                     debug!("{:?}", payload);
 
-                    if let Some(response) = handle_packet(remote_addr, payload, tracker).await {
+                    if let Some(response) = request_handler.handle_packet(remote_addr, payload, tracker).await {
                         UdpServer::send_response(socket, remote_addr, response).await;
                     }
                 }
